requirements.txt
Django==2.1.11
gunicorn
django-heroku
django-bootstrap4
social-auth-app-django
django-widget-tweaks
whitenoise
django-friendship
boto3
django-storages
Pillow
channels
channels_redis
django-model-utils
dj-database-url
psycopg2-binary
pandas
fuzzywuzzy
python-Levenshtein
numpy
scipy
django-webpack-loader
django-cors-headers
python-decouple